added entity relations and other models

+++ Migrations/AppDbContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace OnlineRetailStoreV01.Migrations
 
             modelBuilder.Entity("OnlineRetailStoreV01.Models.User", b =>
                 {
-                    b.Property<long>("UserId")
+                    b.Property<int>("UserId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("UserId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("UserId"));
 
                     b.Property<string>("Email")
                         .IsRequired()
